refactor(#2456): split frankenstein_editor.py into model/panels/coordinator
- Extract URDFModel dataclass to _frankenstein_model.py (249 LOC)
- Extract ModelPanel + StealComponentDialog to _frankenstein_panels.py (427 LOC)
- Slim frankenstein_editor.py coordinator to 550 LOC (was 1173 LOC)
- All public names re-exported for backward compatibility
- TDD contract tests added (9 tests, all GREEN)
